internal/dag: move some builder tests into different files
Moves a few test functions from builder_test.go into
different test files to better align with what they're
testing.
